Display the version badge on crate list pages
Instead of the plain-text version number.

This has the advantage of quickly showing pre-1.0 (orange) vs. post-1.0
(blue) when scanning down a list of crates.

Since people might want to copy the version number of a crate from the
crate list page, I've changed the alt text of the image to just be the
version number that used to appear here, so that you can highlight the
image and copy the version number.