Add 7-day moving averages to download graph. Plus UI tweaks.
The non-stacked scattered chart as a basis reflects the download numbers
correctly. The stacked area chart before was a bit midleading. Both
approaches are a bit "noisy", so I added a 7-day moving average line for
each version number. This gives a higher-level overview and also makes it
easy to see when new versions overtake older ones in usage.

The colors were chosen to be colorblindness-friendly as well as having
a hot-cold metaphor.