Implement request extensions on `T: Request` not `Request`
There are a bunch of tests which are explicitly casting `&MockRequest`
into a `&Request` trait object, because the extension methods provided
are only on the trait object, not anything which implements the trait.

By changing from `Request + 'a` to `T: Request + ?Sized` (which is more
ideomatic anyway), we can just call the method directly on `MockRequest`
in tests.