Follow the Diesel
Get it? Because this ports the following endpoints. To Diesel.
...Anyone?

This ports 4 endpoints over to Diesel. The 3 endpoints which manipulate
the `following` endpoint, as well as the `/me/updates` endpoint since it
is only hit by the tests for the following endpoints.

I ended up changing the updates endpoint quite a bit. I wanted to
eliminate the N+1 queries on the max version, and was wondering why we
needed the max version at all here. I went to go look at it in the UI,
and it turns out that the dashboard page which displayed it is actually
broken as well. After fixing it, I noticed that it doesn't need the
crates at all, just the name (which we tell Ember is the id).

I couldn't actually find a good way in Ember to reference the ID of an
association without loading the whole thing. If anybody knows a better
way to do it than what I'm doing here, please let me know.

Since we don't need the crates, I've just opted not to include that data
in the response body (note that just not including the max version is a
bad idea, since ember caches stuff and it could result in a page that
does need the max version displaying wrong later).

While I was touching these endpoints, I also went ahead and reduced them
all to a single query.

Fixes #438.