Don't tie the lifetime of the db connection to the request
Consider code like the following:

```rust
let conn = req.db_conn()?;
let body = req.body();
```

Since the `body` function requires `&mut self`, this code will fail
since the lifetime of the connection is tied to the lifetime of the
request. These are of course disjoint, but the compiler can't know that.

Really there's no reason to stick the request in `extensions` like we
are. I think the reason the postgres code did it that way was to allow
mocking the connection out in tests, but we just set the pool size to 1.

Additionally, some endpoints (like `crates#summary` are embarassingly
parallelizable, so we should expose the capability for multiple database
connections in a single request.