Port /crates/new and /crates/show to use Diesel
This is probably the single largest endpoint in the entire app. It had a
bit more churn than I'd like, since we're now explicitly wrapping the
whole thing in a database transaction. I've also ported over the
`/crates/show` endpoint as there were several tests that needed both to
be ported over.

There were a large number of tests that needed even more endpoints
ported over in order to be valid. These tests are important, but I want
to keep this commit at least somewhat manageble so I've set them to
ignore for now.