Ensure Diesel NotFound results in a 404
Any method in Diesel's API which returns a single record (specifically
`.get_result` and `.first`) treats 0 records as an error case by
default (calling `.optional` on the `Result` turns it into a
Result<Option<T>>`). However, right now if a
`diesel::result::Error::NotFound` is returned, the server will return a
500. We don't test every possible place a 404 can be returned, so this
is never apparent in tests.

The implementation is a bit of a kludge, but I'm not sure that there's a
better way to do this without either requiring nightly for
specialization or significantly reworking the error handling of the
application (I do think that we should eventually do the latter)