Drop the old schema_migrations table
We haven't been using this since we switched to diesel;
`__diesel_schema_migrations` manages this now. We left this around just
in case we needed to roll back the switch to diesel in production, but
we're good and switched now.

The existence checks are so that these migrations won't error if people
are starting from a database that either does or does not have a
schema_migrations table.