Migrate test database at build time instead of runtime
While we're running these in the integration tests, they weren't run
before unit tests. Rather than duplicating the logic for this everywhere
that we have a test that interacts with the database, we can just do
this at compile time.

As best I can tell, Cargo doesn't give you any environment variable to
know that you're building for tests and not just a debug build, so I've
set this up to only try to migrate if the env var is present, and not
error in its absense.

Fixes #762.