Fix crate/downloads
I appear to have fucked up in #746. The API was sending the detailed
stats for the oldest 5 versions instead of the newest 5 versions,
causing the whole graph to just show "other". This fixes that error.

(as an aside, I kinda prefer just having the whole picture as one stat
like that)