Port `bin/delete-crate` over to Diesel
I moved nearly all of the behavior from this binary into the database.
We could instead just do exactly what this binary was doing before, but
it's a lot of behavior that has 0 tests. I wanted to figure out how best
to actually test this as well, since any test which actually attempted
to delete the crate would have to insert rows in all of these tables to
be useable, meaning it would have to know about every foreign key that
exists. That test would be brittle, and likely to get out of date.

By moving this into the database, we can also test that any new columns
which point to the crates table have a foreign key, and won't break in
the future.

Additionally, I've tested the updated binary locally against an old dump
of the database.

If we're concerned about someone "accidentally" deleting a crate (and we
really shouldn't be. Nobody is running "delete from crates" against the
production database, and we have backups), I'd recommend that we just
have a trigger that checks some
`yes_id_really_like_to_delete_this_crate` entry, or leave exactly 1
constraint uncascading (probably the one on `versions.crate_id`)