Fix keywords that weren't lowercased before inserting
This fixes invalid data that currently exists in the database. Between
commits d5c563560 and 4c2bd4225, it was possible for keywords to be
created without lowercasing them first. Then, due to this code:

```
pub fn find_or_create_all(conn: &PgConnection, names: &[&str]) ->
QueryResult<Vec<Keyword>> {
    // ...elided...

    keywords::table
        .filter(::lower(keywords::keyword).eq(any(&lowercase_names)))
        .load(conn)
}
```

which selected any keywords that matched the given keywords *regardless*
of case and associated these keywords with crates, many crates now have
BOTH the uppercase and lowercase variants of a keyword associated with
them.

That is, the keywords table contains:

| id | keyword |
|----|---------|
| 1  | foo     |
| 2  | Foo     | <-- should not exist

And a crate with id 5 that was published with any of:

`keywords = ["foo"]`
`keywords = ["Foo"]`
`keywords = ["FOO"]` etc

in their Cargo.toml metadata would get the following records created in
the crates_keywords table:

| crate_id | keyword_id |
|----------|------------|
| 5        | 1          | <-- correct
| 5        | 2          | <-- incorrect, for all cases

So this migration removes all the crates_keywords rows associated with
keywords that aren't lowercased and then removes the keywords
themselves.

BUT!

There are some crates that only have these associations currently with
the uppercased keyword! That is, crate id 5 has this in crates_keywords:

| crate_id | keyword_id |
|----------|------------|
| 5        | 2          |

So BEFORE making the fix above, this migration inserts the corresponding
lowercased keyword association for these crates (so 5, 1 in this case)
so that these crates can be treated like all the rest that are
associated with both.

ALSO!

There are some keywords that have only ever been used in their uppercase
variant, so the corresponding lowercase keyword doesn't exist at all.
Such as:

| id | keyword |
|----|---------|
| 3  | BAR     |

In this case, the migration updates the keyword value, leaving the
associated crates_keywords alone.

WHEW!