Switch to `DELETE` for /logout route to prevent CSRF
Other sites are allowed to make `GET` requests so switchig to `DELETE`
to match the destructive nature of the route.

If the user navigates to `GET /logout` we now fall back to the default
behavior which is to serve up the ember app.  Ember then sends a
`DELETE` request.  Fortunately we already enforce sameorigin for
iframes via the `X-FRAME-OPTIONS` header.

Fixes: #986