Switch to `cargo clippy` which also lints the binaries and tests
This switches our CI to install `clippy` and use the `cargo clippy`
command.  Additionally, the tests are now run in the same build as
`rustfmt`.  If `cargo fmt` fails, travis should still run
`cargo clippy` and report any errors there.

In this configuration, `clippy` is no longer included in the
`Cargo.toml` file.