Update to use Diesel 0.99.0
Most of the code changes are for us to switch to `sql_query` where
appropriate. `sql` deprecated its API for bind parameters, as it's meant
to be used with fragments of SQL as part of the query builder, not full
queries. `sql_query` does not require the SQL type to be specified, and
it gets the values by name instead of by index. Unfortunately, none of
the places we are using raw SQL were places that
`#[derive(QueryableByName)]` would work, so I had to manually implement
all of them

This also updates to r2d2 0.8, because I couldn't remember the command
to make cargo force r2d2-diesel to resolve to r2d2 0.7