fix(templates/application): Logo should not be focusable
When using tab to navigate between focusable elements, it does not make sense
to have two adjacent links pointing to the same address. Another solution
would be to place the img tag inside the second address tag, but that leads
to some style issue so I have opted for this approach instead.