Don't require a valid email address to sign in
Users do not necessarily have a valid email in GitHub. We should not
fail most actions if we cannot successfully send them an email. I have
purposely left the error case on the `users/email/token/regenerate`
endpoint, since that endpoint clearly has an email purpose in mind.

I opted to change the signature of the existing name in order to force a
compiler error on any code that hasn't been looked at to determine
whether it should fail on this or not.

Fixes #1142