Don't use shields.io's crate badges
Every single time someone loads a crates.io badge from shields.io, they
make a separate API call to our servers for each one, with no caching.
This means that right now when you go to the search page, you're going
to make 11 API calls when you should be making 1. We have all the
information we need for this badge when we load the page, so let's just
do that and use their "badge with these words" API instead.

I've just removed it from the version show page since we already display
that info elsewhere on the page, and it's especially confusing when the
numbers don't match.