Explicitly never select categories.path column
This is necessary because:
- The path column is of type Ltree
- To get text out of an ltree column, it has to be selected using the
postgres function ltree2text
- We can't modify the select statement generated by the table! macro to
select columns with type ltree differently
- So we'd have to explicitly select columns for all Category queries
anyway
- But we never use categories.path, only query against it, so save all
this trouble and just never select it in the same way Crate never
selects the fulltext search columns.

There might be better ways to do this, or at the very least a better way
of defining the type aliases, but this is what I got working.