Remove part of a test that isn't testing actual logic
This test, which logs in, logs out, then tries to publish a crate with
an invalid token, isn't actually testing useful logic because the cookie
logic for the tests is injected into the test app. Actual browser cookie
logic isn't exercised in these tests; at best, this test is testing the
tests.

The other parts of this test function are fine.