Master changed VersionBuilder and CrateBuilder to
be part of their own builders.rs file and this PR
accidentally merged them back into the all.rs file
when trying to resolve the merge conflict. This
commit removes them from all.rs so that it reflects
master, and then makes the minimal changes needed to
the models in builders.rs to reflect the new database
updates.