Log the state of the connection pool once per minute
On rare occasions we've had outages occur as a result of the connection
pool refusing to return any connections. Shortly before this occurs, we
start to see significant spikes in response times (likely because the
pool has begun serving only a 1 or 2 connections and there's a large
amount of contention).

Right now we have limited capability to debug this. The only
explanations that I can come up with are a bug in r2d2, or something is
happening causing our worker threads to be killed in a way that does not
run destructors, or something is causing threads to hang indefinitely.

If one of those three options above is true, we should see `connections
- idle_connections` trend towards zero over time, and diverge away from
`in_flight_requests` over time (they aren't expected to be exactly
equal, since some requests will never attempt to get a database
connection, and we can have more requests than available connections
during normal operations)

This isn't going to be enough information to definitively figure out
what's going on here, but it should hopefully give us *some* information
to help figure out where to look next