Use a consistent format for numbers in dl graph
When formatting download numbers in the download graph, use a consistent
numbering format for both raw numbers and the rolling 7-day average.

There are cases where the user's local can make inconsistent formatting.
Specifically, we use `Number.prototype.toFixed` to format the rolling
7-day average, which always results in a number format that uses a
decimal point (e.g. "1234.56"). However, the raw download numbers end up
using the locale, which in some places can end up formatting the number
with a decimal as the thousands separator ("1.234").

To fix this, use google.visualization.NumberFormat for *both* values, so
that they are always consistent.